rename DocUrl.ts to AutomergeUrl.ts
